Fix timeout should be unsigned long, not int.

+++ include/gyros/cond.h
@@ -39,7 +39,8 @@ void gyros_cond_init(gyros_cond_t *c);
 
 void gyros_cond_wait(gyros_cond_t *c, gyros_mutex_t *m);
 
-int gyros_cond_timedwait(gyros_cond_t *c, gyros_mutex_t *m, int timeout);
+int gyros_cond_timedwait(gyros_cond_t *c, gyros_mutex_t *m,
+                         unsigned long timeout);
 
 void gyros_cond_signal_one(gyros_cond_t *c);
